test(tls): keep the duplex-teardown cases in one file and assert the whole sequence
The pending-handshake case only excluded ECONNRESET. Assert the exact
close sequence like the other cases do.

node-tls-connect.test.ts already owns the TLS-over-Duplex cases, both
client and server wraps, so the server cases move there too.
