+++ test/js/node/tls/node-tls-server.test.ts
@@ -5,7 +5,6 @@ import https from "https";
 import net, { AddressInfo } from "net";
 import { createTest } from "node-harness";
 import { once } from "node:events";
-import stream from "node:stream";
 import { tmpdir } from "os";
 import { join } from "path";
 import type { PeerCertificate } from "tls";
@@ -2032,95 +2031,6 @@ it("destroys a server wrap whose socket was destroyed before the deferred upgrad
     conn?.destroy();
     rawServer.close();
   }
-});
-
-describe("a server wrap over a generic Duplex follows the transport's teardown", () => {
-  // Node wraps a stream that has no fd in a JSStreamSocket and destroys the
-  // TLS socket when that wrap closes:
-  // https://github.com/nodejs/node/blob/v26.3.0/lib/internal/tls/wrap.js#L739-L741
-  // So a destroyed transport always reaches the TLS socket as 'close' with
-  // hadError === false. Bun runs a TLS engine over the stream instead, and
-  // that engine is created on a later event-loop turn, so the cases below
-  // cover both sides of that turn.
-  const makeTransport = () =>
-    new stream.Duplex({
-      read() {},
-      write(_chunk, _encoding, callback) {
-        callback();
-      },
-    });
-  // Resolves with the events the socket emitted, in order.
-  const recordTeardown = (socket: TLSSocket) => {
-    const events: string[] = [];
-    const closed = Promise.withResolvers<string[]>();
-    socket.on("error", (err: Error & { code?: string }) => events.push(`error:${err.code}`));
-    socket.on("close", hadError => {
-      events.push(`close:${hadError}`);
-      closed.resolve(events);
-    });
-    return closed.promise;
-  };
-
-  it("destroying the transport in the same tick as the wrap closes the TLS socket", async () => {
-    // The engine does not exist yet, so the close has to be held and reported
-    // when it starts. Without that the socket never closes at all.
-    const transport = makeTransport();
-    const wrapped = new TLSSocket(transport, { isServer: true, ...COMMON_CERT });
-    const teardown = recordTeardown(wrapped);
-    transport.destroy();
-    expect(await teardown).toEqual(["close:false"]);
-    expect(wrapped.destroyed).toBe(true);
-  });
-
-  it("destroying the transport after the engine started closes the TLS socket", async () => {
-    // The engine is up and its handshake is still pending. The aborted
-    // handshake must not surface as an ECONNRESET instead of the close.
-    const transport = makeTransport();
-    const wrapped = new TLSSocket(transport, { isServer: true, ...COMMON_CERT });
-    const teardown = recordTeardown(wrapped);
-    // 'secure' never fires without a peer, but the upgrade task that creates
-    // the engine is queued ahead of this one.
-    await new Promise<void>(resolve => setImmediate(resolve));
-    transport.destroy();
-    expect(await teardown).not.toContain("error:ECONNRESET");
-    expect(wrapped.destroyed).toBe(true);
-  });
-
-  it("destroying the transport of an established connection closes the TLS socket", async () => {
-    // Each side's _write pushes into the other, so this handshakes for real
-    // over two in-memory duplexes (what Playwright's client-certificate proxy
-    // does with its browser-facing transport).
-    const makeSide = (peer: () => stream.Duplex) =>
-      new stream.Duplex({
-        read() {},
-        write(chunk, _encoding, callback) {
-          peer().push(chunk);
-          callback();
-        },
-        final(callback) {
-          peer().push(null);
-          callback();
-        },
-      });
-    const clientSide: stream.Duplex = makeSide(() => serverSide);
-    const serverSide: stream.Duplex = makeSide(() => clientSide);
-    const wrapped = new TLSSocket(serverSide, { isServer: true, ...COMMON_CERT });
-    const client = connect({ socket: clientSide, rejectUnauthorized: false });
-    try {
-      const secured = Promise.withResolvers<void>();
-      client.on("error", secured.reject);
-      wrapped.on("error", secured.reject);
-      wrapped.on("secure", () => secured.resolve());
-      await secured.promise;
-      const teardown = recordTeardown(wrapped);
-      serverSide.destroy();
-      expect(await teardown).toEqual(["close:false"]);
-      expect(wrapped.destroyed).toBe(true);
-    } finally {
-      client.destroy();
-      wrapped.destroy();
-    }
-  });
 });
 
 it("exposes the server-side peer verification result via socket.ssl.verifyError()", async () => {
